fixed int to long when cast the 404

+++ LisaApplication/src/project/cs/lisa/application/DownloadWebPageTask.java
@@ -88,7 +88,7 @@ public class DownloadWebPageTask extends AsyncTask<URL, Void, Void> {
                     JSONObject searchResult = (JSONObject) obj;
 
                     // get from uplink
-                    if (searchResult == null || ((Integer) searchResult.get("status")) == 404) {
+                    if (searchResult == null || ((Long) searchResult.get("status")) == 404) {
                         downloadAndDisplayWebPage();
                         return;
                     }
